requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
PyJWT==1.6.4
pytz==2018.4
requests==2.19.1
urllib3==1.23
